split up attribute collection for meta and time index. significantly reduced mem use since number of spatial chunks tends to be much lower than number of time chunks.

+++ sup3r/postprocessing/writers/base.py
@@ -381,7 +381,7 @@ class OutputHandler(OutputMixin):
             mins.append(min_val)
 
         data = np.maximum(data, mins)
-        return np.minimum(data, maxes)
+        return np.minimum(data, maxes).astype(np.float32)
 
     @staticmethod
     def pad_lat_lon(lat_lon):
